feat: Resolve SASS decprecation warnings
## Description
The release of SASS 3.0.0 will change a number of behaviours within SASS
code. There are a number of warnings currently being output to highlight
changes that are required in order to keep up-to-date.

This change makes those update to resolve the highlighted issues and also
updates our dependencies to esnure they are also up-to-date.

## Issue
https://jira.dev.bbc.co.uk/browse/GEL